requirements.txt
python-telegram-bot==21.9
fastapi==0.115.0
uvicorn==0.30.6
